Use argparse, instead of hacky command-line parsing.
My code to parse command lines did the job, but argparse does it
better.  Adding new commands will be much easier now.
For instance, the command handlers don't need to parse arguments
any more, since argparse does it.  They are much simpler.

argparse is included in Python 2.7, but it is fairly easy to get for
Python 2.6.