Add "smart" repository detection.
Some commands will have an optional --repository argument.
If --repository is not supplied, then we try to automatically detect
the repository.
If the current working directory is part of a repo cloned from github,
then the name of the current repository is found by reading
remote.origin.url, using the "git config" command.