Fix README.
The README gave the git:// URL for cloning; change to https.
Probably my original reason for that choice was my allergy to the
web and the HTTP-ification of everything.  Why clone over http
when git has a wire protocol and URI scheme of its own?
But the reality is that it isn't encrypted, while https is.
If gits:// (git over TLS) ever becomes a thing, I'll
update the README yet again.